requirements.txt
Flask==2.1.3
flask-restx==0.5.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
marshmallow==3.16.0
PyJWT==2.4.0
flask_cors
werkzeug==2.1.2
